Move slot code to its own crate
This was highly non-trivial, due to cyclic dependencies.
