+++ Cargo.toml
@@ -27,6 +27,7 @@ members = [
 	"core/consensus/aura",
 	"core/consensus/babe",
 	"core/consensus/rhd",
+	"core/consensus/slots",
 	"core/executor",
 	"core/finality-grandpa",
 	"core/finality-grandpa/primitives",
