+++ core/consensus/aura/Cargo.toml
@@ -13,7 +13,7 @@ runtime_support = { package = "srml-support", path = "../../../srml/support" }
 runtime_primitives = { package = "sr-primitives", path = "../../sr-primitives" }
 runtime_version = { package = "sr-version", path = "../../sr-version" }
 runtime_io = { package = "sr-io", path = "../../sr-io" }
-aura_slots = { package = "substrate-consensus-aura-slots", path = "slots" }
+slots = { package = "substrate-consensus-slots", path = "../slots" }
 aura_primitives = { package = "substrate-consensus-aura-primitives", path = "primitives" }
 inherents = { package = "substrate-inherents", path = "../../inherents" }
 srml-consensus = { path = "../../../srml/consensus" }
