+++ core/consensus/slots/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "substrate-consensus-slots"
+version = "1.0.0"
+authors = ["Parity Technologies <admin@parity.io>"]
+description = "Generic slots-based utilities for consensus"
+edition = "2018"
+
+[dependencies]
+codec = { package = "parity-codec", version = "3.2" }
+client = { package = "substrate-client", path = "../../client" }
+primitives = { package = "substrate-primitives", path = "../../primitives" }
+runtime_primitives = { package = "sr-primitives", path = "../../sr-primitives" }
+consensus_common = { package = "substrate-consensus-common", path = "../common" }
+inherents = { package = "substrate-inherents", path = "../../inherents" }
+futures = "0.1.17"
+tokio = "0.1.7"
+parking_lot = "0.7.1"
+error-chain = "0.12"
+log = "0.4"
