+++ core/test-runtime/Cargo.toml
@@ -15,6 +15,7 @@ substrate-client = { path = "../client", default-features = false }
 primitives = { package = "substrate-primitives", path = "../primitives", default-features = false }
 inherents = { package = "substrate-inherents", path = "../inherents", default-features = false }
 consensus_aura = { package = "substrate-consensus-aura-primitives", path = "../consensus/aura/primitives", default-features = false }
+slots = { package = "substrate-consensus-slots", path = "../consensus/slots", default-features = false }
 rstd = { package = "sr-std", path = "../sr-std", default-features = false }
 runtime_io = { package = "sr-io", path = "../sr-io", default-features = false }
 runtime_primitives = { package = "sr-primitives", path = "../sr-primitives", default-features = false }
